implementing the ext calling conventions for Apple EABI.
This is a little bit tricky, as these conventions do not directly fit
to the framework (which doesn't support register pairs).
The workaround is that we pretend that the first two float args are
passed via d0 and d1, and that the single float result comes back in
d0. This is not the truth, but fits into the framework. In reality,
the float args go into r0/r1, r1/r2 or r2/r3, and are mixed with any
int or pointer args. The float result is in r0/r1. When we emit
the call instruction, we also emit additional code that moves
the floats from d0 and d1 to the register pairs (and potentially
also moves the integers away that are already there). After
returning from the call, the float, if any, is loaded into d0.
The price of this workaround: up to three additional moves per
extcall.